🚚 Rename to plural
